Handle pdf download in export modal
