+++ front/app/containers/Admin/communityMonitor/CommunityMonitorFormBuilder/utils.tsx
@@ -3,6 +3,7 @@ import React from 'react';
 import { Box } from '@citizenlab/cl2-component-library';
 
 import AccessRightsNotice from 'containers/Admin/projects/project/nativeSurvey/AccessRightsNotice';
+import UserFieldsInFormNotice from 'containers/Admin/projects/project/nativeSurvey/UserFieldsInFormNotice';
 
 import { FormBuilderConfig } from 'components/FormBuilder/utils';
 import Warning from 'components/UI/Warning';
@@ -10,7 +11,6 @@ import Warning from 'components/UI/Warning';
 import { FormattedMessage } from 'utils/cl-intl';
 
 import messages from './messages';
-import UserFieldsInFormNotice from 'containers/Admin/projects/project/nativeSurvey/UserFieldsInFormNotice';
 
 export const communityMonitorConfig: FormBuilderConfig = {
   type: 'survey',
