+++ front/app/containers/Admin/projects/project/nativeSurvey/SurveyFormBuilder/index.tsx
@@ -6,13 +6,8 @@ import useFormCustomFields from 'api/custom_fields/useCustomFields';
 import usePhase from 'api/phases/usePhase';
 import useProjectById from 'api/projects/useProjectById';
 
-import useLocale from 'hooks/useLocale';
-
-import { FormPDFExportFormValues } from 'containers/Admin/projects/components/PDFExportModal';
-
 import FormBuilder from 'components/FormBuilder/edit';
 
-import { saveSurveyAsPDF } from '../saveSurveyAsPDF';
 import { nativeSurveyConfig, clearOptionAndStatementIds } from '../utils';
 
 const SurveyFormBuilder = ({
@@ -27,7 +22,6 @@ const SurveyFormBuilder = ({
   const { data: phase } = usePhase(phaseId);
   const { data: project } = useProjectById(projectId);
 
-  const locale = useLocale();
   const { data: formCustomFields } = useFormCustomFields({
     projectId,
     phaseId: copyFrom ? copyFrom : phaseId,
@@ -44,20 +38,12 @@ const SurveyFormBuilder = ({
     ? formCustomFields
     : clearOptionAndStatementIds(formCustomFields);
 
-  // PDF downloading
-  const handleExportPDF = async ({
-    personal_data,
-  }: FormPDFExportFormValues) => {
-    await saveSurveyAsPDF({ phaseId, locale, personal_data });
-  };
-
   return (
     <FormBuilder
       builderConfig={{
         ...nativeSurveyConfig,
         formCustomFields: newCustomFields,
         goBackUrl: `/admin/projects/${projectId}/phases/${phaseId}/native-survey`,
-        onDownloadPDF: handleExportPDF,
       }}
       viewFormLink={`/projects/${project.data.attributes.slug}/surveys/new?phase_id=${phase.data.id}`}
     />
